requirements.txt
Flask==2.2.2
catboost==1.2
azure-storage-blob==2.1.0
Flask-Cors==4.0.0